requirements.txt
kfp
matplotlib
scikit-learn
numpy
joblib
pandas
google-cloud-storage
google-cloud-bigquery
tensorflow